Refactored AzLogin
